requirements.txt
django==4.1.2
mysqlclient==2.1.1
pymysql==1.0.2
djangorestframework==3.14.0 
django-cors-headers==3.13.0